requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.0

# ML and Vector Processing
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.3.2

# Privacy and Security
cryptography==41.0.8
pynacl==1.5.0
differential-privacy==1.1.5

# Database and Storage
redis==5.0.1
elasticsearch==8.11.0
psycopg2-binary==2.9.9

# Async and Networking
aiohttp==3.9.1
asyncio-mqtt==0.16.1
grpcio==1.60.0
grpcio-tools==1.60.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
opentelemetry-api==1.21.0

# Configuration and CLI
click==8.1.7
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2